Revert "Pwa 2.0"
