+++ public/index.html
@@ -10,7 +10,7 @@
           content="Agentz Messenger">
     <title>Agentz Messenger</title>
     <!-- For removing IOS bounce effect -->
-    <!-- <script defer src="https://cdnjs.cloudflare.com/ajax/libs/inobounce/0.2.0/inobounce.min.js"></script> -->
+    <script defer src="https://cdnjs.cloudflare.com/ajax/libs/inobounce/0.2.0/inobounce.min.js"></script>
     <link rel="stylesheet" href="loader.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/jvectormap/2.0.4/jquery-jvectormap.css" type="text/css" media="screen"/>
     <meta name="apple-mobile-web-app-title" content="Agentz">
@@ -61,9 +61,6 @@
 
         
       </script>
-      <noscript>
-        <iframe  frameborder="0" style="height: 100vh; width: 100vw;"></iframe> 
-      </noscript>
 </head>
 
 <body id="body">
