+++ src/app/routes/chatPanel/redux/index.js
@@ -72,11 +72,11 @@ class ChatPanelWithRedux extends PureComponent {
     this.ChangeUrl('/app/chat/'+user['id']);
     if(document.getElementById('selectedUser')){
       var div = document.getElementById('selectedUser');
-        div.innerHTML = user.name || user.emailId || user.contactNo;
+        div.innerHTML = user.name || user.emailId || this.formatPhoneNumber(user.contactNo);
     }
     if(document.getElementById('selectedContactNo')){
       var div = document.getElementById('selectedContactNo');
-      let contactNo = user.contactNo ? user.contactNo : '';
+      let contactNo = user.contactNo ? this.formatPhoneNumber(user.contactNo) : '';
       div.innerHTML =  user.name || user.emailId ? contactNo ? contactNo : '' : '';
     }
     if(document.getElementById('phone')){
@@ -94,6 +94,35 @@ class ChatPanelWithRedux extends PureComponent {
     } else {
         alert("Browser does not support HTML5.");
     }
+  }
+
+
+   formatPhoneNumber = (inputStr) => {
+    //Filter only numbers from the input
+    let returnString = inputStr;
+    if(!inputStr.startsWith('+') && inputStr.length > 10){
+      inputStr = '+' + inputStr;
+    }
+    let returnStr = parsePhoneNumberFromString(inputStr);
+    let phoneNumberString = inputStr;
+    let areaCode = null;
+    if(returnStr){
+     areaCode = returnStr.countryCallingCode ? returnStr.countryCallingCode : null ;
+      phoneNumberString = returnStr.nationalNumber;
+    }else{
+      areaCode = '1';
+    }
+     var cleaned = ('' + phoneNumberString).replace(/\D/g, '')
+     var match = cleaned.match(/^(\d{3})(\d{3})(\d{4})$/)
+     if (match) {
+      returnString = '';
+      if(areaCode){
+        returnString = '+' + areaCode;
+      }
+       returnString =  returnString + ' (' + match[1] + ') ' + match[2] + '-' + match[3];
+     }
+    
+     return returnString;
   }
 
   submitComment = () => {
